frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = "Streamlit application for IPID document parsing"
authors = ["Robin Beaudet <robin.beaudet@zelros.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
python-multipart = "^0.0.5"
streamlit = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
